Background Scroller
I couldn't figure it out completely because the way other tutorials are
done have the player fixated...my solution was to fixate the background
to the camera and move the background to the left using scroll
speed...then after a certain amount, it would tile back using the 2
background Quads.  I feel like for an infinite runner though, it
wouldn't get past 2 or 3 of these backgrounds at the current speed.